:recycle: Improve forging enabling Provide users an ability to use the forging status Warn user before overwrite the forging information Refactored forging enable and disable by removing base forging
